First Commit for Problem.java and Answer.java
This is the Problem Class for CSCI-466 Networks Project 1,
Trivia Game Server.  This class will take the answers from a database of

questions/answers and creates an object for each one.  The 'getter' for
the answers will scramble the order of the answers to ensure we do not
have patterns developing that would allow the players to deduce the
answer
based on their relative position on the display.